+++ src/main/java/lagatrix/server/manager/OSManager.java
@@ -1,10 +1,12 @@
 package lagatrix.server.manager;
 
 import lagatrix.server.entities.dto.os.OSInformation;
+import lagatrix.server.exceptions.NotValidFamilyException;
 import lagatrix.server.exceptions.manager.os.OSException;
 import lagatrix.server.manager.information.os.DistributionInfo;
 import lagatrix.server.manager.information.os.UnameInfo;
 import lagatrix.server.tools.command.CommandExecutor;
+import lagatrix.server.tools.detectors.PackageManagerDetector;
 
 /**
  * This class obtain information of the OS of Linux.
@@ -32,8 +34,9 @@ public class OSManager {
      *
      * @return The OS information entity.
      * @throws OSException If you can't get some information from OS.
+     * @throws NotValidFamilyException If the family of OS is not supported.
      */
-    public OSInformation obtainOSInformation() throws OSException {
+    public OSInformation obtainOSInformation() throws OSException, NotValidFamilyException {
         OSInformation os = new OSInformation();
 
         os.setDistribution(informationDistribution.obtainDistribution());
@@ -42,6 +45,7 @@ public class OSManager {
         os.setDistributionVersion(informationDistribution.obtainDistributionVersion());
         os.setHostname(informationUname.obtainHostname());
         os.setKernel(informationUname.obtainKernel());
+        os.setPackageManager(PackageManagerDetector.detectPackageManager(os.getDistributionFamily()));
 
         return os;
     }